Convert to Gradle app; Convert configs to support Java >= 11
Lots of content here. Since splitting this up into multiple commits would be more confusing than seeing the changes in-bulk (due to how many of the changes are interrelated), this will be a god-commit, but counter-balanced with good code comments (which would be used when diagnosing issues, using this repo as a boilerplate for a new repo, etc.) since comments would be seen and adhered to before commit messages would anyway.